changed to use size_t lengths

+++ Source/sysex.c
@@ -3,11 +3,10 @@
 /* convert to/from sysex 7-bit data 
  * taken from http://blogs.bl0rg.net/netzstaub/2008/08/14/encoding-8-bit-data-in-midi-sysex/
  */
-uint8_t data_to_sysex(uint8_t *data, uint8_t *sysex, uint8_t len) {
-  uint8_t retlen = 0;
-  uint8_t cnt;
+size_t data_to_sysex(uint8_t *data, uint8_t *sysex, size_t len) {
+  size_t retlen = 0;
+  size_t cnt;
   uint8_t cnt7 = 0;
-
   sysex[0] = 0;
   for(cnt = 0; cnt < len; cnt++) {
     uint8_t c = data[cnt] & 0x7F;
@@ -24,9 +23,9 @@ uint8_t data_to_sysex(uint8_t *data, uint8_t *sysex, uint8_t len) {
   return retlen + cnt7 + (cnt7 != 0 ? 1 : 0);
 }
 
-uint8_t sysex_to_data(uint8_t *sysex, uint8_t *data, uint8_t len) {
-  uint8_t cnt;
-  uint8_t cnt2 = 0;
+size_t sysex_to_data(uint8_t *sysex, uint8_t *data, size_t len) {
+  size_t cnt;
+  size_t cnt2 = 0;
   uint8_t bits = 0;
   for(cnt = 0; cnt < len; cnt++) {
     if((cnt % 8) == 0) {
